Add files via upload
To use this script, save it to a file with a .sh extension and make it executable by running chmod +x script.sh. Then, you can run the script by typing ./script.sh in the terminal.

Note: This script will only work if you are running it on a Windows machine with the Windows Subsystem for Linux (WSL) installed, and if you have the systemctl and sc commands available.

It's important to note that disabling Windows Defender may leave your machine vulnerable to malware and other security threats. It is generally not recommended to disable Windows Defender unless you have another antivirus program installed and configured to provide protection for your system.